+++ src/components/Login/method/TokenLogin.tsx
@@ -3,31 +3,35 @@ import React, { useEffect } from 'react';
 import useMutation, { IMutation } from 'src/libs/client/useMutation';
 import { LoginForm } from '../components/LoginForm';
 
-export const UserLogin = ({ method }: any) => {
-  const [login, { loading, data }] = useMutation<IMutation>(
-    '/api/users/user_login',
-  );
+export interface TokenForm {
+  token: string;
+}
 
-  console.log(data);
+function TokenLogin({ method }: any) {
+  const [login, { loading, data }] = useMutation<IMutation>(
+    '/api/users/token_login',
+  );
   const errMsg = data?.errorMsg;
 
-  //페이지이동
+  //페이지 이동
   const router = useRouter();
   useEffect(() => {
     if (data?.ok) {
       router.push('/my-page');
     }
   }, [data, router]);
-
   //
+
   return (
     <>
       <LoginForm
         method={method}
+        errMsg={errMsg}
         login={login}
         loading={loading}
-        errMsg={errMsg}
       />
     </>
   );
-};
+}
+
+export default TokenLogin;
