+++ src/components/Login/method/UserLogin.tsx
@@ -3,35 +3,29 @@ import React, { useEffect } from 'react';
 import useMutation, { IMutation } from 'src/libs/client/useMutation';
 import { LoginForm } from '../components/LoginForm';
 
-export interface TokenForm {
-  token: string;
-}
+export const UserLogin = ({ method }: any) => {
+  const [login, { loading, data }] = useMutation<IMutation>('/api/users/login');
 
-function TokenLogin({ method }: any) {
-  const [login, { loading, data }] = useMutation<IMutation>(
-    '/api/users/token_session',
-  );
+  console.log(data);
   const errMsg = data?.errorMsg;
 
-  //페이지 이동
+  //페이지이동
   const router = useRouter();
   useEffect(() => {
     if (data?.ok) {
       router.push('/my-page');
     }
   }, [data, router]);
-  //
 
+  //
   return (
     <>
       <LoginForm
         method={method}
-        errMsg={errMsg}
         login={login}
         loading={loading}
+        errMsg={errMsg}
       />
     </>
   );
-}
-
-export default TokenLogin;
+};
