+++ src/libs/client/useUser.ts
@@ -1,6 +1,6 @@
 import useSWR from 'swr';
 
-export default function useLoggedIn() {
+export default function useUser() {
   const { data, error } = useSWR('/api/users/profile');
   return {
     loggedIn: data?.ok,
