+++ src/pages/login.tsx
@@ -1,6 +1,6 @@
 import { Container, H1 } from 'src/styles/loginStyles';
 import { LoginLink } from 'src/components/Login/components/LoginLink';
-import { AuthLogin } from 'src/components/Login/LoginMethod/AuthLogin';
+import { AuthLogin } from 'src/components/Login/method/AuthLogin';
 
 export default function Login() {
   return (
